BLD: Change build setup closer to ldf (to fix clean)
Try to do mall cleanups to the build process, mainly started
because `build.sh clean` was now broken.
This isn't enough... the build seems messy to me, most likely
the Python and C++ cmake files should be split?
(Or maybe we should just not bother with building the C++
library at all in it's own step.)

As is, Python rebuilds everything, it seems...
(And because the cuda-arches are weird locally, it is very slow)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>