remove cython dependency, remove setup.cfg (#10)
Came here looking to remove the `setup.cfg` file, inspired by
https://github.com/rapidsai/raft/pull/2532.

But this PR uncovered a few other packaging things to be cleaned up.
Proposes:

* removing `Cython` dependency
* aligning `cmake_minimum_required()` version with the one from the
conda recipe, and with what RAPIDS has been using for the last few
releases
* removing `maptlolib` dependency for benchmarks

## Notes for Reviewers

### How is it safe to remove `Cython`?

As far as I can tell, this project doesn't have any Cython code. It is
building a shared library and then loading it with `ctypes`:


https://github.com/rapidsai/legate-raft/blob/3260952df5cc09fe356ccf07e8832070f1029abb/legate_raft/library.py#L21-L26

Signed-off-by: James Lamb <jlamb@nvidia.com>