MAINT: Make changes to one-pool and introduce simple mapper (#11)
The mapper is basically taken from legate-boost, we always should have
had one probably (for the C-order thing).

Otherwise add with-allocations where needed (it might be that I missed a
task, but then it is probably unused and should be deleted).

Also deleted some files that were clearly unused (the parts they belong
to should be slashed, not revived).

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>