Bump to legate 25.05 and maintenance as needed (#19)
To be honest, I don't like the task changes needed much. There is a
small chance we can revert almost all of them again.
There is also a chance we have to re-organize things more, but for now
get to building with 25.05...

(Local connectivity issues, maybe abuse CI a bit more than usually...)

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>